XMLHttpRequest/tests/submissions/Opera/send-send.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - send()</title>
    <script src="/resources/testharness.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/delay.php?ms=0")
        client.send(null)
        assert_throws("INVALID_STATE_ERR", function() { client.send(null) })
        client.abort()
      })
    </script>
  </body>
</html>
